works for me

+++ go/loop.go
@@ -21,11 +21,11 @@
 package fibo
 
 // Fast fibonacci calculation
-func Fast(n int32) int32 {
+func Fast(n int64) int64 {
 	if n <= 1 {
 		return 1
 	}
-	var x, p, pp int32
+	var x, p, pp int64
 	pp, p = 0, 1
 	for x = 2; x <= n; x++ {
 		pp, p = p, p + pp
